backend/requirements.txt
#backend/requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.20
bcrypt==4.2.1